Updated Layout.
This commit updates the layout of the GUI. It changes it from a absolute positioning to using a box layout. It also adds the ability to create a set of randomized plots and edit them (change color and style).

The Plot class is updated to store the different data plotted so that it can be accessed to change things for that specific data. As of right now, it cannot remove any plots.